Allow optimizing for more than one class
I've played around with strategies on how to balance multiple classes
against each other. If you just use an "n hot" encoding, one of the
classes tends to receive a lot of weight while the others starve.
In the current version, if a desired class a is activated with k, and the
weakest desired activation is l, I set the desired activation of a to
l/k (within bounds). Problem: This gives a far lower highest activation.
Also, it's entirely unclear whether 30 tabby cat and 30 granny smith
really make a balanced apple-cat.